HDDS-6489: Fix Unit test issue
